leaks ?

+++ src/crm/gobelins/facedroidtactoe/GameConsts.java
@@ -3,10 +3,10 @@ package crm.gobelins.facedroidtactoe;
 public class GameConsts {
 
 	public static final int GAME_WIDTH = 3;
-	public static final double LEFT = 0.175;
-	public static final double RIGHT = 0.875;
-	public static final double TOP = 0.25;
-	public static final double BOTTOM = 0.5;
-	public static final double WIDTH = 0.7;
-	public static final double HEIGHT = 0.25;
+	public static final float LEFT = (float) 0.175;
+	public static final float RIGHT = (float) 0.875;
+	public static final float TOP = (float) 0.25;
+	public static final float BOTTOM = (float) 0.5;
+	public static final float WIDTH = (float) 0.7;
+	public static final float HEIGHT = (float) 0.25;
 }
